move eligibility flag to dropdown
